backend/requirements.txt
livekit-agents[openai,silero]~=1.0
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn